Move formatting records to a separate module
To ease in testing.
